==> src/module/step8_calculate_result.py <==
"""
Created by Chungkang Choi
March 2024

Description: Calculate MCA Result
"""

import pandas as pd
import rasterio
import numpy as np
import copy

def process_result_calculation(input_path, output_path, input_excel_path):
    # create panda data frame for each purpose
    df_input_excel = pd.read_excel(input_excel_path)

    # Filter rows where are AOI, exclusion
    scored_df = df_input_excel[(df_input_excel['AOI'] != 1) | (df_input_excel['exclusion'] != 1)]

    # Calculate 'layer_weight_rate' column
    df_input_excel['layer_weight_rate'] = (df_input_excel['layer_weight'] / scored_df['layer_weight'].sum()).round(3)

    # Modifying file_name to include "AOI"
    # AOI data frame
    df_AOI = df_input_excel[df_input_excel['AOI'] == 1]
    AOI_file_path = input_path + df_AOI.iloc[0]['file_name']

    # create base raster array for raster calculation, should be 0 value / using for raster cell calculation
    with rasterio.open(AOI_file_path) as AOI_dataset:
        array_calculation = np.array(copy.deepcopy(AOI_dataset.read(1).astype(float)))-1

    # filtering scored layers: layer_weight is not null
    df_scored_layers = df_input_excel[df_input_excel['layer_weight_rate'].notna()]
    
    # filtering non-scored layers: layer_weight is null (AOI, exclusion)
    df_nonscored_layers = df_input_excel[df_input_excel['layer_weight_rate'].isna()]

    # calculate layer weight
    for index, row in df_scored_layers.iterrows():
        file_path = input_path + row['file_name']
       
        try:
            with rasterio.open(file_path) as dataset:
                dataset_raster_value = dataset.read(1)
                # Checking for NoData values and setting them to 0
                if dataset.nodata is not None:
                    dataset_raster_value[dataset_raster_value == dataset.nodata] = 0

                # Verifying if the raster data type matches the expected type
                if not np.issubdtype(dataset_raster_value.dtype, np.number):
                    print(f"Error: Raster data type is not numeric at path {file_path}")
                    continue

                # Performing array operations
                array_calculation += dataset_raster_value * row['layer_weight_rate']

        except rasterio.errors.RasterioIOError as e:
            print(f"Error reading raster file at path {file_path}: {e}")

    # calculate AOI, Exclusion
    for index, row in df_nonscored_layers.iterrows():
        # Access values in each row
        file_path = input_path + row['file_name']
        include_AOI = row['AOI']
        exclude = row['exclusion']

        if include_AOI == 1:  # multiply
            with rasterio.open(file_path) as dataset:
                dataset_raster_value = dataset.read(1)
                array_calculation *= dataset_raster_value

        elif exclude == 1:  # 0,1 inverted multiply
            with rasterio.open(file_path) as dataset:
                dataset_raster_value = dataset.read(1) # read raster data
                no_data_value = dataset.nodata # Check NoData value

                # Setting cells with NoData to 0
                if no_data_value is not None:
                    dataset_raster_value[dataset_raster_value == no_data_value] = 0

                inverted_dataset_raster_value = 1 - dataset_raster_value
                array_calculation *= inverted_dataset_raster_value

    # smaller than 1 means NoData
    no_data_value = 0
    array_calculation[array_calculation < 1] = no_data_value

    with rasterio.open(AOI_file_path) as AOI_dataset:
        profile = AOI_dataset.profile
        profile.update(dtype=rasterio.float32, nodata=no_data_value, compress='DEFLATE')
    
    with rasterio.open(output_path + r'MCA_result.tif', 'w', **profile) as dst:
        dst.write(array_calculation.astype(profile['dtype']), 1)


def main():
    # set input file path
    input_path = r'data\\step7\\'
    output_path = r'data\\step8\\'
    setting_excel_path = r'data\\setting_excel\\'
    input_excel_path = setting_excel_path + r'step7_excel_template.xlsx'
    process_result_calculation(input_path, output_path, input_excel_path)

if __name__ == "__main__":
    main()
